avtka: resizing of ui enabled
This commit introduces basic resizing of the whole drawn UI.
It relies on cairo_scale() to perform the heavy lifting of
mapping the drawn lines/etc to the device space.

The math of finding the best resize-ratio to match the original
("native") width and height, compared to the resized width/height
is performed in the draw call. The ratio is fixed - so the aspect
ratio of the window must remain constant (or black parts will show).

Finally, button and mouse events are scaled by the inverse of the
drawing scale - to "unmap" the scale, so that the widget bounding
boxes can be handled in the original "native" UI draw size.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>