meson: don't install library if a subproject
Previously, if another Meson project used AVTKA as a subproject,
AVTKA would install it self if the ninja install step was run.
This isn't desired, as on the next iteration of "meson configure",
the other project will pick up the now "installed" version of AVTKA,
instead of the dep. This also causes it to not checkout the latest
git version, which causes headaches.

Hopefully solved by setting install to false if AVTKA is a subproject.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>