screen: initial implementation of virtual screen
This commit adds a new concept to the AVTKA library, allowing
a "screen" to be created. The screen represents a physical
screen on a controller, allowing a developer to prototype and
test the integration with the device's screen.

Currently there are still some hard-coded constants and
assumptions in the code. These will remain until a decision
is made on how to present the screens.

There are two main choices:
1) Application draws in a known format (ARGB32 or somthing). Apps can
   use eg: Cairo, and the Ctlra library accepts a pixel array, which
   it transforms into the binary stream the device expects.
2) The Application gets raw-access to the binary stream, and must decode
   or otherwise fixup the correct byte layout for the device.

There is a compromise available, where the raw pointer to the data is
exposed, but Ctlra *ALSO* provides a helper function that just accepts
RGB24 data, and converts to the data-format as required by the hw instance.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>