avtka: enable off-screen buffer for Ctlra
This commit enables an "offscreen" mode, where the buffer being
used by AVKTA is not backed by PUGL or X11, but is an AVTKA
created Cairo surface. This enables "embedded" use-cases where
no UI is required or desired. An example is redrawing a Ctlra
screen seperately to the X11 vsync / refresh callback - which
is required when X11 windows get hidden as they don't redraw.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>