screen: fix scaling and outline screen on UI
This commit improves the scale and position of screen
drawing in AVTKA. The sizes are now taken from the
application, and the redraw size is not the pixel size,
but instead the "screen pixel surface" is rescaled before
being drawn to the main UI canvas.

As a result, the screen should show in the correct position
and have the correct relative size to the other items on the
UI, however the contents of the screen are re-scaled.

Future work can improve screen-debugging by providing a method
to screen-shot the sub-screen-surface, allowing developers to
see the quality of thier work natively.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>